Move cargo's self-version code to a dedicated module.
